Remove: 불필요한 파일삭제 - Detail 컴포넌트 삭제
