--- src/pages/ProductDetail/components/Detail.scss
@@ -1,19 +1,0 @@
-@import '../../../styles/variable.scss';
-
-.Detail {
-  width: 100vw;
-  margin-top: calc($padding * 3);
-  padding: calc($padding * 5);
-  background-color: $white-color;
-
-  .tab-contents {
-    @include flexSort(space-evenly, flex-start);
-    padding: $padding;
-    border-bottom: 2px solid $grey-color;
-
-    a {
-      text-decoration: none;
-      color: $green-dark-color;
-    }
-  }
-}
